finish adjusting the route and controllers

+++ app/Http/Controllers/EmployeeController.php
@@ -1,89 +1,53 @@
 <?php
 
 namespace App\Http\Controllers;
-use App\Events\NewEmployeeRegisteredEvent;
-use App\Models\employee;
-use App\Http\Requests\employeesvalid;
-use Illuminate\Support\Facades\Gate;
-use Illuminate\Http\Request;
-use App\Jobs\welcomejob;
 
+use App\Http\Requests\employeesvalid;
+use App\Models\employee;
 
 class EmployeeController extends Controller
 {
+    public function index()
+    {
+        $employee = $employee = employee::all();
 
-
-  public function index()
-  {
-    $employee = $employee = employee::all();
-    return view('employee.index', ['employee' => $employee]);
-  }
-
-
-
-  public function store(employeesvalid $request)
-  {
-
-    if (Gate::denies('isUser')) {
-    $request->user()->employees()->create($request->validated());
-     // event(new NewEmployeeRegisteredEvent($employee));
-     $details = "serge";
-     welcomejob::dispatch($details);
-      return redirect('/employee')->with('msgs', 'successfully updated');
-
-
-    } else {
-      return redirect('/error');
+        return view('employee.index', ['employee' => $employee]);
     }
 
+    public function store(employeesvalid $request)
+    {
 
+        $this->authorize('create', employee::class);
+        $request->user()->employees()->create($request->validated());
 
-  }
-
-
-
-
-  public function salary(employee $employee)
-  {
-
-    return view('test.delete', ['employ' => $employee]);
-  }
-
-
-
-
-  public function edit(employee $employee)
-  {
-    
-     return view('employee.edite', ['employee' => $employee]);
-  }
-
-
-  public function update(employeesvalid $request, employee $employee)
-  {
-
-    if (!Gate::allows('isUser')) {
-      
-      $employee->update($request->validated());
-
-      return redirect('/employee')->with('msgs', 'successfully updated');
-
-
-    } else {
-      return redirect('/error');
+        return redirect('/employee')->with('msgs', 'successfully updated');
     }
 
+    public function salary(employee $employee)
+    {
 
+        return view('test.delete', ['employ' => $employee]);
+    }
 
-  }
+    public function edit(employee $employee)
+    {
 
+        return view('employee.edite', ['employee' => $employee]);
+    }
 
+    public function update(employeesvalid $request, employee $employee)
+    {
+        $this->authorize('update', $employee);
+        $employee->update($request->validated());
 
-  public function destroy(employee $employee)
-  {
-    $employee->delete();
-    return redirect()->route('employee.index');
-  }
+        return redirect('/employee')->with('msgs', 'successfully updated');
+    }
 
+    public function destroy(employee $employee)
+    {
+        $this->authorize('delete', $employee);
+        $employee->delete();
 
+        return redirect()->route('employee.index');
+    }
 }
